Organize test code
