--- src/converters/index.ts
@@ -1,6 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT License.
-
-export * from './BindingConverters';
-export * from './RpcConverters';
-export * from './RpcHttpConverters';
